Split eventtrace into separate files and remove ETW::TypeSystemLog
