+++ src/coreclr/nativeaot/Runtime/eventtrace.cpp
@@ -13,414 +13,14 @@
 
 #include "daccess.h"
 
-#include "slist.h"
-#include "varint.h"
-#include "regdisplay.h"
-#include "StackFrameIterator.h"
 #include "thread.h"
 #include "threadstore.h"
 #include "threadstore.inl"
-//#include "PalRedhawk.h"
-#include "EventPipeInterface.h"
-
-#define Win32EventWrite PalEventWrite
-
-#ifdef FEATURE_ETW
-#include "eventtracepriv.h"
-#endif // FEATURE_ETW
 
 volatile LONGLONG ETW::GCLog::s_l64LastClientSequenceNumber = 0;
 
-/****************************************************************************/
-/* Methods that are called from the runtime */
-/****************************************************************************/
-
-
-// Simple helpers called by the GC to decide whether it needs to do a walk of heap
-// objects and / or roots.
-
-#ifdef FEATURE_ETW
-
-BOOL ETW::GCLog::ShouldWalkHeapObjectsForEtw()
-{
-    LIMITED_METHOD_CONTRACT;
-    return RUNTIME_PROVIDER_CATEGORY_ENABLED(
-        TRACE_LEVEL_INFORMATION,
-        CLR_GCHEAPDUMP_KEYWORD);
-}
-
-BOOL ETW::GCLog::ShouldWalkHeapRootsForEtw()
-{
-    LIMITED_METHOD_CONTRACT;
-    return RUNTIME_PROVIDER_CATEGORY_ENABLED(
-        TRACE_LEVEL_INFORMATION,
-        CLR_GCHEAPDUMP_KEYWORD);
-}
-
-BOOL ETW::GCLog::ShouldTrackMovementForEtw()
-{
-    LIMITED_METHOD_CONTRACT;
-    return RUNTIME_PROVIDER_CATEGORY_ENABLED(
-        TRACE_LEVEL_INFORMATION,
-        CLR_GCHEAPSURVIVALANDMOVEMENT_KEYWORD);
-}
-
-BOOL ETW::GCLog::ShouldWalkStaticsAndCOMForEtw()
-{
-    // @TODO:
-    return FALSE;
-}
-
-void ETW::GCLog::WalkStaticsAndCOMForETW()
-{
-    // @TODO:
-}
-
-
-// Batches the list of moved/surviving references for the GCBulkMovedObjectRanges /
-// GCBulkSurvivingObjectRanges events
-struct EtwGcMovementContext
-{
-public:
-    // An instance of EtwGcMovementContext is dynamically allocated and stored
-    // inside of MovedReferenceContextForEtwAndProfapi, which in turn is dynamically
-    // allocated and pointed to by a profiling_context pointer created by the GC on the stack.
-    // This is used to batch and send GCBulkSurvivingObjectRanges events and
-    // GCBulkMovedObjectRanges events. This method is passed a pointer to
-    // MovedReferenceContextForEtwAndProfapi::pctxEtw; if non-NULL it gets returned;
-    // else, a new EtwGcMovementContext is allocated, stored in that pointer, and
-    // then returned. Callers should test for NULL, which can be returned if out of
-    // memory
-    static EtwGcMovementContext* GetOrCreateInGCContext(EtwGcMovementContext** ppContext)
-    {
-        LIMITED_METHOD_CONTRACT;
-
-        _ASSERTE(ppContext != NULL);
-
-        EtwGcMovementContext* pContext = *ppContext;
-        if (pContext == NULL)
-        {
-            pContext = new (nothrow) EtwGcMovementContext;
-            *ppContext = pContext;
-        }
-        return pContext;
-    }
-
-    EtwGcMovementContext() :
-        iCurBulkSurvivingObjectRanges(0),
-        iCurBulkMovedObjectRanges(0)
-    {
-        LIMITED_METHOD_CONTRACT;
-        Clear();
-    }
-
-    // Resets structure for reuse on construction, and after each flush.
-    // (Intentionally leave iCurBulk* as is, since they persist across flushes within a GC.)
-    void Clear()
-    {
-        LIMITED_METHOD_CONTRACT;
-        cBulkSurvivingObjectRanges = 0;
-        cBulkMovedObjectRanges = 0;
-        ZeroMemory(rgGCBulkSurvivingObjectRanges, sizeof(rgGCBulkSurvivingObjectRanges));
-        ZeroMemory(rgGCBulkMovedObjectRanges, sizeof(rgGCBulkMovedObjectRanges));
-    }
-
-    //---------------------------------------------------------------------------------------
-    // GCBulkSurvivingObjectRanges
-    //---------------------------------------------------------------------------------------
-
-    // Sequence number for each GCBulkSurvivingObjectRanges event
-    UINT iCurBulkSurvivingObjectRanges;
-
-    // Number of surviving object ranges currently filled out in rgGCBulkSurvivingObjectRanges array
-    UINT cBulkSurvivingObjectRanges;
-
-    // Struct array containing the primary data for each GCBulkSurvivingObjectRanges
-    // event. Fix the size so the total event stays well below the 64K limit (leaving
-    // lots of room for non-struct fields that come before the values data)
-    EventStructGCBulkSurvivingObjectRangesValue rgGCBulkSurvivingObjectRanges[
-        (cbMaxEtwEvent - 0x100) / sizeof(EventStructGCBulkSurvivingObjectRangesValue)];
-
-    //---------------------------------------------------------------------------------------
-    // GCBulkMovedObjectRanges
-    //---------------------------------------------------------------------------------------
-
-    // Sequence number for each GCBulkMovedObjectRanges event
-    UINT iCurBulkMovedObjectRanges;
-
-    // Number of Moved object ranges currently filled out in rgGCBulkMovedObjectRanges array
-    UINT cBulkMovedObjectRanges;
-
-    // Struct array containing the primary data for each GCBulkMovedObjectRanges
-    // event. Fix the size so the total event stays well below the 64K limit (leaving
-    // lots of room for non-struct fields that come before the values data)
-    EventStructGCBulkMovedObjectRangesValue rgGCBulkMovedObjectRanges[
-        (cbMaxEtwEvent - 0x100) / sizeof(EventStructGCBulkMovedObjectRangesValue)];
-};
-
-// Contains above struct for ETW, plus extra info (opaque to us) used by the profiling
-// API to track its own information.
-struct MovedReferenceContextForEtwAndProfapi
-{
-    // An instance of MovedReferenceContextForEtwAndProfapi is dynamically allocated and
-    // pointed to by a profiling_context pointer created by the GC on the stack. This is used to
-    // batch and send GCBulkSurvivingObjectRanges events and GCBulkMovedObjectRanges
-    // events and the corresponding callbacks for profapi profilers. This method is
-    // passed a pointer to a MovedReferenceContextForEtwAndProfapi; if non-NULL it gets
-    // returned; else, a new MovedReferenceContextForEtwAndProfapi is allocated, stored
-    // in that pointer, and then returned. Callers should test for NULL, which can be
-    // returned if out of memory
-    static MovedReferenceContextForEtwAndProfapi* CreateInGCContext(LPVOID pvContext)
-    {
-        LIMITED_METHOD_CONTRACT;
-
-        _ASSERTE(pvContext != NULL);
-
-        MovedReferenceContextForEtwAndProfapi* pContext = *(MovedReferenceContextForEtwAndProfapi**)pvContext;
-
-        // Shouldn't be called if the context was already created.  Perhaps someone made
-        // one too many BeginMovedReferences calls, or didn't have an EndMovedReferences
-        // in between?
-        _ASSERTE(pContext == NULL);
-
-        pContext = new (nothrow) MovedReferenceContextForEtwAndProfapi;
-        *(MovedReferenceContextForEtwAndProfapi**)pvContext = pContext;
-
-        return pContext;
-    }
-
-
-    MovedReferenceContextForEtwAndProfapi() :
-        pctxProfAPI(NULL),
-        pctxEtw(NULL)
-
-    {
-        LIMITED_METHOD_CONTRACT;
-    }
-
-    LPVOID pctxProfAPI;
-    EtwGcMovementContext* pctxEtw;
-};
-
-
-//---------------------------------------------------------------------------------------
-//
-// Called by the GC for each moved or surviving reference that it encounters. This
-// batches the info into our context's buffer, and flushes that buffer to ETW as it fills
-// up.
-//
-// Arguments:
-//      * pbMemBlockStart - Start of moved/surviving block
-//      * pbMemBlockEnd - Next pointer after end of moved/surviving block
-//      * cbRelocDistance - How far did the block move? (0 for non-compacted / surviving
-//          references; negative if moved to earlier addresses)
-//      * profilingContext - Where our context is stored
-//      * fCompacting - Is this a compacting GC? Used to decide whether to send the moved
-//          or surviving event
-//
-
-// static
-void ETW::GCLog::MovedReference(
-    BYTE* pbMemBlockStart,
-    BYTE* pbMemBlockEnd,
-    ptrdiff_t cbRelocDistance,
-    size_t profilingContext,
-    BOOL fCompacting,
-    BOOL /*fAllowProfApiNotification*/) // @TODO: unused param from newer implementation
-{
-    CONTRACTL
-    {
-        NOTHROW;
-        GC_NOTRIGGER;
-        MODE_ANY;
-        CAN_TAKE_LOCK;  // EEToProfInterfaceImpl::AllocateMovedReferencesData takes lock
-    }
-    CONTRACTL_END;
-
-    MovedReferenceContextForEtwAndProfapi* pCtxForEtwAndProfapi =
-        (MovedReferenceContextForEtwAndProfapi*)profilingContext;
-    if (pCtxForEtwAndProfapi == NULL)
-    {
-        _ASSERTE(!"MovedReference() encountered a NULL profilingContext");
-        return;
-    }
-
-#ifdef PROFILING_SUPPORTED
-    // ProfAPI
-    {
-        BEGIN_PIN_PROFILER(CORProfilerTrackGC());
-        g_profControlBlock.pProfInterface->MovedReference(pbMemBlockStart,
-            pbMemBlockEnd,
-            cbRelocDistance,
-            &(pCtxForEtwAndProfapi->pctxProfAPI),
-            fCompacting);
-        END_PIN_PROFILER();
-    }
-#endif // PROFILING_SUPPORTED
-
-    // ETW
-
-    if (!ShouldTrackMovementForEtw())
-        return;
-
-    EtwGcMovementContext* pContext =
-        EtwGcMovementContext::GetOrCreateInGCContext(&pCtxForEtwAndProfapi->pctxEtw);
-    if (pContext == NULL)
-        return;
-
-    if (fCompacting)
-    {
-        // Moved references
-
-        _ASSERTE(pContext->cBulkMovedObjectRanges < _countof(pContext->rgGCBulkMovedObjectRanges));
-        EventStructGCBulkMovedObjectRangesValue* pValue =
-            &pContext->rgGCBulkMovedObjectRanges[pContext->cBulkMovedObjectRanges];
-        pValue->OldRangeBase = pbMemBlockStart;
-        pValue->NewRangeBase = pbMemBlockStart + cbRelocDistance;
-        pValue->RangeLength = pbMemBlockEnd - pbMemBlockStart;
-        pContext->cBulkMovedObjectRanges++;
-
-        // If buffer is now full, empty it into ETW
-        if (pContext->cBulkMovedObjectRanges == _countof(pContext->rgGCBulkMovedObjectRanges))
-        {
-            FireEtwGCBulkMovedObjectRanges(
-                pContext->iCurBulkMovedObjectRanges,
-                pContext->cBulkMovedObjectRanges,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGCBulkMovedObjectRanges[0]),
-                &pContext->rgGCBulkMovedObjectRanges[0]);
-
-            pContext->iCurBulkMovedObjectRanges++;
-            pContext->Clear();
-        }
-    }
-    else
-    {
-        // Surviving references
-
-        _ASSERTE(pContext->cBulkSurvivingObjectRanges < _countof(pContext->rgGCBulkSurvivingObjectRanges));
-        EventStructGCBulkSurvivingObjectRangesValue* pValue =
-            &pContext->rgGCBulkSurvivingObjectRanges[pContext->cBulkSurvivingObjectRanges];
-        pValue->RangeBase = pbMemBlockStart;
-        pValue->RangeLength = pbMemBlockEnd - pbMemBlockStart;
-        pContext->cBulkSurvivingObjectRanges++;
-
-        // If buffer is now full, empty it into ETW
-        if (pContext->cBulkSurvivingObjectRanges == _countof(pContext->rgGCBulkSurvivingObjectRanges))
-        {
-            FireEtwGCBulkSurvivingObjectRanges(
-                pContext->iCurBulkSurvivingObjectRanges,
-                pContext->cBulkSurvivingObjectRanges,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGCBulkSurvivingObjectRanges[0]),
-                &pContext->rgGCBulkSurvivingObjectRanges[0]);
-
-            pContext->iCurBulkSurvivingObjectRanges++;
-            pContext->Clear();
-        }
-    }
-}
-
-
-//---------------------------------------------------------------------------------------
-//
-// Called by the GC just before it begins enumerating plugs.  Gives us a chance to
-// allocate our context structure, to allow us to batch plugs before firing events
-// for them
-//
-// Arguments:
-//      * pProfilingContext - Points to location on stack (in GC function) where we can
-//         store a pointer to the context we allocate
-//
-
-// static
-void ETW::GCLog::BeginMovedReferences(size_t* pProfilingContext)
-{
-    LIMITED_METHOD_CONTRACT;
-
-    MovedReferenceContextForEtwAndProfapi::CreateInGCContext(LPVOID(pProfilingContext));
-}
-
-
-//---------------------------------------------------------------------------------------
-//
-// Called by the GC at the end of a heap walk to give us a place to flush any remaining
-// buffers of data to ETW or the profapi profiler
-//
-// Arguments:
-//      profilingContext - Our context we built up during the heap walk
-//
-
-// static
-void ETW::GCLog::EndMovedReferences(size_t profilingContext,
-    BOOL /*fAllowProfApiNotification*/) // @TODO: unused param from newer implementation
-{
-    CONTRACTL
-    {
-        NOTHROW;
-        GC_NOTRIGGER;
-        MODE_ANY;
-        CAN_TAKE_LOCK;
-    }
-    CONTRACTL_END;
-
-    MovedReferenceContextForEtwAndProfapi* pCtxForEtwAndProfapi = (MovedReferenceContextForEtwAndProfapi*)profilingContext;
-    if (pCtxForEtwAndProfapi == NULL)
-    {
-        _ASSERTE(!"EndMovedReferences() encountered a NULL profilingContext");
-        return;
-    }
-
-#ifdef PROFILING_SUPPORTED
-    // ProfAPI
-    {
-        BEGIN_PIN_PROFILER(CORProfilerTrackGC());
-        g_profControlBlock.pProfInterface->EndMovedReferences(&(pCtxForEtwAndProfapi->pctxProfAPI));
-        END_PIN_PROFILER();
-    }
-#endif //PROFILING_SUPPORTED
-
-    // ETW
-
-    if (!ShouldTrackMovementForEtw())
-        return;
-
-    // If context isn't already set up for us, then we haven't been collecting any data
-    // for ETW events.
-    EtwGcMovementContext* pContext = pCtxForEtwAndProfapi->pctxEtw;
-    if (pContext == NULL)
-        return;
-
-    // Flush any remaining moved or surviving range data
-
-    if (pContext->cBulkMovedObjectRanges > 0)
-    {
-        FireEtwGCBulkMovedObjectRanges(
-            pContext->iCurBulkMovedObjectRanges,
-            pContext->cBulkMovedObjectRanges,
-            GetClrInstanceId(),
-            sizeof(pContext->rgGCBulkMovedObjectRanges[0]),
-            &pContext->rgGCBulkMovedObjectRanges[0]);
-    }
-
-    if (pContext->cBulkSurvivingObjectRanges > 0)
-    {
-        FireEtwGCBulkSurvivingObjectRanges(
-            pContext->iCurBulkSurvivingObjectRanges,
-            pContext->cBulkSurvivingObjectRanges,
-            GetClrInstanceId(),
-            sizeof(pContext->rgGCBulkSurvivingObjectRanges[0]),
-            &pContext->rgGCBulkSurvivingObjectRanges[0]);
-    }
-
-    pCtxForEtwAndProfapi->pctxEtw = NULL;
-    delete pContext;
-}
-
-/***************************************************************************/
-/* This implements the public runtime provider's ManagedHeapCollectKeyword.  It
-   performs a full, gen-2, blocking GC.
-/***************************************************************************/
+// This implements the public runtime provider's GCHeapCollectKeyword.  It
+// performs a full, gen-2, blocking GC.
 void ETW::GCLog::ForceGC(LONGLONG l64ClientSequenceNumber)
 {
     CONTRACTL
@@ -444,8 +44,6 @@ void ETW::GCLog::ForceGC(LONGLONG l64ClientSequenceNumber)
 
     ForceGCForDiagnostics();
 }
-
-#endif // FEATURE_ETW
 
 //---------------------------------------------------------------------------------------
 //
@@ -479,7 +77,6 @@ void ETW::GCLog::FireGcStart(ETW_GC_INFO* pGcInfo)
     }
 }
 
-#ifdef FEATURE_ETW
 //---------------------------------------------------------------------------------------
 //
 // Contains code common to profapi and ETW scenarios where the profiler wants to force
@@ -530,610 +127,12 @@ HRESULT ETW::GCLog::ForceGCForDiagnostics()
     return hr;
 }
 
-
-//---------------------------------------------------------------------------------------
-// BulkTypeValue / BulkTypeEventLogger: These take care of batching up types so they can
-// be logged via ETW in bulk
-//---------------------------------------------------------------------------------------
-
-BulkTypeValue::BulkTypeValue()
-    : cTypeParameters(0)
-    , rgTypeParameters()
-    , ullSingleTypeParameter(0)
+COOP_PINVOKE_HELPER(void, RhpEtwExceptionThrown, (LPCWSTR exceptionTypeName, LPCWSTR exceptionMessage, void* faultingIP, HRESULT hresult))
 {
-    LIMITED_METHOD_CONTRACT;
-    ZeroMemory(&fixedSizedData, sizeof(fixedSizedData));
+    FireEtXplatExceptionThrown_V1(exceptionTypeName,
+        exceptionMessage,
+        faultingIP,
+        hresult,
+        0,
+        GetClrInstanceId());
 }
-
-//---------------------------------------------------------------------------------------
-//
-// Clears a BulkTypeValue so it can be reused after the buffer is flushed to ETW
-//
-
-void BulkTypeValue::Clear()
-{
-    CONTRACTL
-    {
-        THROWS;
-        GC_NOTRIGGER;
-        MODE_ANY;
-    }
-    CONTRACTL_END;
-
-    ZeroMemory(&fixedSizedData, sizeof(fixedSizedData));
-    cTypeParameters = 0;
-    ullSingleTypeParameter = 0;
-    rgTypeParameters.Release();
-}
-
-//---------------------------------------------------------------------------------------
-//
-// Fire an ETW event for all the types we batched so far, and then reset our state
-// so we can start batching new types at the beginning of the array.
-//
-//
-
-void BulkTypeEventLogger::FireBulkTypeEvent()
-{
-    LIMITED_METHOD_CONTRACT;
-
-    if (m_nBulkTypeValueCount == 0)
-    {
-        // No types were batched up, so nothing to send
-        return;
-    }
-
-    // Normally, we'd use the MC-generated FireEtwBulkType for all this gunk, but
-    // it's insufficient as the bulk type event is too complex (arrays of structs of
-    // varying size). So we directly log the event via EventDataDescCreate and
-    // EventWrite
-
-    // We use one descriptor for the count + one for the ClrInstanceID + 4
-    // per batched type (to include fixed-size data + name + param count + param
-    // array).  But the system limit of 128 descriptors per event kicks in way
-    // before the 64K event size limit, and we already limit our batch size
-    // (m_nBulkTypeValueCount) to stay within the 128 descriptor limit.
-    EVENT_DATA_DESCRIPTOR EventData[128];
-    UINT16 nClrInstanceID = GetClrInstanceId();
-
-    UINT iDesc = 0;
-
-    _ASSERTE(iDesc < _countof(EventData));
-    EventDataDescCreate(&EventData[iDesc++], &m_nBulkTypeValueCount, sizeof(m_nBulkTypeValueCount));
-
-    _ASSERTE(iDesc < _countof(EventData));
-    EventDataDescCreate(&EventData[iDesc++], &nClrInstanceID, sizeof(nClrInstanceID));
-
-    for (int iTypeData = 0; iTypeData < m_nBulkTypeValueCount; iTypeData++)
-    {
-        // Do fixed-size data as one bulk copy
-        _ASSERTE(iDesc < _countof(EventData));
-        EventDataDescCreate(
-            &EventData[iDesc++],
-            &(m_rgBulkTypeValues[iTypeData].fixedSizedData),
-            sizeof(m_rgBulkTypeValues[iTypeData].fixedSizedData));
-
-        // Do var-sized data individually per field
-
-        // Type name (nonexistent and thus empty on nativeaot)
-        _ASSERTE(iDesc < _countof(EventData));
-        EventDataDescCreate(&EventData[iDesc++], L"", sizeof(WCHAR));
-
-        // Type parameter count
-        _ASSERTE(iDesc < _countof(EventData));
-        EventDataDescCreate(
-            &EventData[iDesc++],
-            &(m_rgBulkTypeValues[iTypeData].cTypeParameters),
-            sizeof(m_rgBulkTypeValues[iTypeData].cTypeParameters));
-
-        // Type parameter array
-        if (m_rgBulkTypeValues[iTypeData].cTypeParameters > 0)
-        {
-            _ASSERTE(iDesc < _countof(EventData));
-            EventDataDescCreate(
-                &EventData[iDesc++],
-                ((m_rgBulkTypeValues[iTypeData].cTypeParameters == 1) ?
-                    &(m_rgBulkTypeValues[iTypeData].ullSingleTypeParameter) :
-                    (ULONGLONG*)(m_rgBulkTypeValues[iTypeData].rgTypeParameters)),
-                sizeof(ULONGLONG) * m_rgBulkTypeValues[iTypeData].cTypeParameters);
-        }
-    }
-
-    Win32EventWrite(Microsoft_Windows_DotNETRuntimeHandle, &BulkType, iDesc, EventData);
-
-    // Reset state
-    m_nBulkTypeValueCount = 0;
-    m_nBulkTypeValueByteCount = 0;
-}
-
-// Holds state that batches of roots, nodes, edges, and types as the GC walks the heap
-// at the end of a collection.
-class EtwGcHeapDumpContext
-{
-public:
-    // An instance of EtwGcHeapDumpContext is dynamically allocated and stored inside of
-    // ProfilingScanContext and ProfilerWalkHeapContext, which are context structures
-    // that the GC heap walker sends back to the callbacks. This method is passed a
-    // pointer to ProfilingScanContext::pvEtwContext or
-    // ProfilerWalkHeapContext::pvEtwContext; if non-NULL it gets returned; else, a new
-    // EtwGcHeapDumpContext is allocated, stored in that pointer, and then returned.
-    // Callers should test for NULL, which can be returned if out of memory
-    static EtwGcHeapDumpContext* GetOrCreateInGCContext(LPVOID* ppvEtwContext)
-    {
-        LIMITED_METHOD_CONTRACT;
-
-        _ASSERTE(ppvEtwContext != NULL);
-
-        EtwGcHeapDumpContext* pContext = (EtwGcHeapDumpContext*)*ppvEtwContext;
-        if (pContext == NULL)
-        {
-            pContext = new (nothrow) EtwGcHeapDumpContext;
-            *ppvEtwContext = pContext;
-        }
-        return pContext;
-    }
-
-    EtwGcHeapDumpContext() :
-        iCurBulkRootEdge(0),
-        iCurBulkRootConditionalWeakTableElementEdge(0),
-        iCurBulkNodeEvent(0),
-        iCurBulkEdgeEvent(0),
-        bulkTypeEventLogger()
-    {
-        LIMITED_METHOD_CONTRACT;
-        ClearRootEdges();
-        ClearRootConditionalWeakTableElementEdges();
-        ClearNodes();
-        ClearEdges();
-    }
-
-    // These helpers clear the individual buffers, for use after a flush and on
-    // construction.  They intentionally leave the indices (iCur*) alone, since they
-    // persist across flushes within a GC
-
-    void ClearRootEdges()
-    {
-        LIMITED_METHOD_CONTRACT;
-        cGcBulkRootEdges = 0;
-        ZeroMemory(rgGcBulkRootEdges, sizeof(rgGcBulkRootEdges));
-    }
-
-    void ClearRootConditionalWeakTableElementEdges()
-    {
-        LIMITED_METHOD_CONTRACT;
-        cGCBulkRootConditionalWeakTableElementEdges = 0;
-        ZeroMemory(rgGCBulkRootConditionalWeakTableElementEdges, sizeof(rgGCBulkRootConditionalWeakTableElementEdges));
-    }
-
-    void ClearNodes()
-    {
-        LIMITED_METHOD_CONTRACT;
-        cGcBulkNodeValues = 0;
-        ZeroMemory(rgGcBulkNodeValues, sizeof(rgGcBulkNodeValues));
-    }
-
-    void ClearEdges()
-    {
-        LIMITED_METHOD_CONTRACT;
-        cGcBulkEdgeValues = 0;
-        ZeroMemory(rgGcBulkEdgeValues, sizeof(rgGcBulkEdgeValues));
-    }
-
-    //---------------------------------------------------------------------------------------
-    // GCBulkRootEdge
-    //
-    // A "root edge" is the relationship between a source "GCRootID" (i.e., stack
-    // variable, handle, static, etc.) and the target "RootedNodeAddress" (the managed
-    // object that gets rooted).
-    //
-    //---------------------------------------------------------------------------------------
-
-    // Sequence number for each GCBulkRootEdge event
-    UINT iCurBulkRootEdge;
-
-    // Number of root edges currently filled out in rgGcBulkRootEdges array
-    UINT cGcBulkRootEdges;
-
-    // Struct array containing the primary data for each GCBulkRootEdge event.  Fix the size so
-    // the total event stays well below the 64K
-    // limit (leaving lots of room for non-struct fields that come before the root edge data)
-    EventStructGCBulkRootEdgeValue rgGcBulkRootEdges[(cbMaxEtwEvent - 0x100) / sizeof(EventStructGCBulkRootEdgeValue)];
-
-
-    //---------------------------------------------------------------------------------------
-    // GCBulkRootConditionalWeakTableElementEdge
-    //
-    // These describe dependent handles, which simulate an edge connecting a key NodeID
-    // to a value NodeID.
-    //
-    //---------------------------------------------------------------------------------------
-
-    // Sequence number for each GCBulkRootConditionalWeakTableElementEdge event
-    UINT iCurBulkRootConditionalWeakTableElementEdge;
-
-    // Number of root edges currently filled out in rgGCBulkRootConditionalWeakTableElementEdges array
-    UINT cGCBulkRootConditionalWeakTableElementEdges;
-
-    // Struct array containing the primary data for each GCBulkRootConditionalWeakTableElementEdge event.  Fix the size so
-    // the total event stays well below the 64K
-    // limit (leaving lots of room for non-struct fields that come before the root edge data)
-    EventStructGCBulkRootConditionalWeakTableElementEdgeValue rgGCBulkRootConditionalWeakTableElementEdges
-        [(cbMaxEtwEvent - 0x100) / sizeof(EventStructGCBulkRootConditionalWeakTableElementEdgeValue)];
-
-    //---------------------------------------------------------------------------------------
-    // GCBulkNode
-    //
-    // A "node" is ANY managed object sitting on the heap, including RootedNodeAddresses
-    // as well as leaf nodes.
-    //
-    //---------------------------------------------------------------------------------------
-
-    // Sequence number for each GCBulkNode event
-    UINT iCurBulkNodeEvent;
-
-    // Number of nodes currently filled out in rgGcBulkNodeValues array
-    UINT cGcBulkNodeValues;
-
-    // Struct array containing the primary data for each GCBulkNode event.  Fix the size so
-    // the total event stays well below the 64K
-    // limit (leaving lots of room for non-struct fields that come before the node data)
-    EventStructGCBulkNodeValue rgGcBulkNodeValues[(cbMaxEtwEvent - 0x100) / sizeof(EventStructGCBulkNodeValue)];
-
-    //---------------------------------------------------------------------------------------
-    // GCBulkEdge
-    //
-    // An "edge" is the relationship between a source node and its referenced target
-    // node. Edges are reported in bulk, separately from Nodes, but it is expected that
-    // the consumer read the Node and Edge streams together. One takes the first node
-    // from the Node stream, and then reads EdgeCount entries in the Edge stream, telling
-    // you all of that Node's targets. Then, one takes the next node in the Node stream,
-    // and reads the next entries in the Edge stream (using this Node's EdgeCount to
-    // determine how many) to find all of its targets. This continues on until the Node
-    // and Edge streams have been fully read.
-    //
-    // GCBulkRootEdges are not duplicated in the GCBulkEdge events. GCBulkEdge events
-    // begin at the GCBulkRootEdge.RootedNodeAddress and move forward.
-    //
-    //---------------------------------------------------------------------------------------
-
-    // Sequence number for each GCBulkEdge event
-    UINT iCurBulkEdgeEvent;
-
-    // Number of nodes currently filled out in rgGcBulkEdgeValues array
-    UINT cGcBulkEdgeValues;
-
-    // Struct array containing the primary data for each GCBulkEdge event.  Fix the size so
-    // the total event stays well below the 64K
-    // limit (leaving lots of room for non-struct fields that come before the edge data)
-    EventStructGCBulkEdgeValue rgGcBulkEdgeValues[(cbMaxEtwEvent - 0x100) / sizeof(EventStructGCBulkEdgeValue)];
-
-
-    //---------------------------------------------------------------------------------------
-    // BulkType
-    //
-    // Types are a bit more complicated to batch up, since their data is of varying
-    // size.  BulkTypeEventLogger takes care of the pesky details for us
-    //---------------------------------------------------------------------------------------
-
-    BulkTypeEventLogger bulkTypeEventLogger;
-};
-
-
-
-//---------------------------------------------------------------------------------------
-//
-// Called during a heap walk for each root reference encountered.  Batches up the root in
-// the ETW context
-//
-// Arguments:
-//      * pvHandle - If the root is a handle, this points to the handle
-//      * pRootedNode - Points to object that is rooted
-//      * pSecondaryNodeForDependentHandle - For dependent handles, this is the
-//          secondary object
-//      * fDependentHandle - nonzero iff this is for a dependent handle
-//      * profilingScanContext - The shared profapi/etw context built up during the heap walk.
-//      * dwGCFlags - Bitmask of "GC_"-style flags set by GC
-//      * rootFlags - Bitmask of EtwGCRootFlags describing the root
-//
-
-// static
-void ETW::GCLog::RootReference(
-    LPVOID pvHandle,
-    Object* pRootedNode,
-    Object* pSecondaryNodeForDependentHandle,
-    BOOL fDependentHandle,
-    ProfilingScanContext* profilingScanContext,
-    DWORD dwGCFlags,
-    DWORD rootFlags)
-{
-    LIMITED_METHOD_CONTRACT;
-
-    if (pRootedNode == NULL)
-        return;
-
-    EtwGcHeapDumpContext* pContext =
-        EtwGcHeapDumpContext::GetOrCreateInGCContext(&profilingScanContext->pvEtwContext);
-    if (pContext == NULL)
-        return;
-
-    // Determine root kind, root ID, and handle-specific flags
-    LPVOID pvRootID = NULL;
-    BYTE nRootKind = (BYTE)profilingScanContext->dwEtwRootKind;
-    switch (nRootKind)
-    {
-    case kEtwGCRootKindStack:
-        break;
-
-    case kEtwGCRootKindHandle:
-        pvRootID = pvHandle;
-        break;
-
-    case kEtwGCRootKindFinalizer:
-        _ASSERTE(pvRootID == NULL);
-        break;
-
-    case kEtwGCRootKindOther:
-    default:
-        _ASSERTE(nRootKind == kEtwGCRootKindOther);
-        _ASSERTE(pvRootID == NULL);
-        break;
-    }
-
-    // Convert GC root flags to ETW root flags
-    if (dwGCFlags & GC_CALL_INTERIOR)
-        rootFlags |= kEtwGCRootFlagsInterior;
-    if (dwGCFlags & GC_CALL_PINNED)
-        rootFlags |= kEtwGCRootFlagsPinning;
-
-    // Add root edge to appropriate buffer
-    if (fDependentHandle)
-    {
-        _ASSERTE(pContext->cGCBulkRootConditionalWeakTableElementEdges <
-            _countof(pContext->rgGCBulkRootConditionalWeakTableElementEdges));
-        EventStructGCBulkRootConditionalWeakTableElementEdgeValue* pRCWTEEdgeValue =
-            &pContext->rgGCBulkRootConditionalWeakTableElementEdges[pContext->cGCBulkRootConditionalWeakTableElementEdges];
-        pRCWTEEdgeValue->GCKeyNodeID = pRootedNode;
-        pRCWTEEdgeValue->GCValueNodeID = pSecondaryNodeForDependentHandle;
-        pRCWTEEdgeValue->GCRootID = pvRootID;
-        pContext->cGCBulkRootConditionalWeakTableElementEdges++;
-
-        // If RCWTE edge buffer is now full, empty it into ETW
-        if (pContext->cGCBulkRootConditionalWeakTableElementEdges ==
-            _countof(pContext->rgGCBulkRootConditionalWeakTableElementEdges))
-        {
-            FireEtwGCBulkRootConditionalWeakTableElementEdge(
-                pContext->iCurBulkRootConditionalWeakTableElementEdge,
-                pContext->cGCBulkRootConditionalWeakTableElementEdges,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGCBulkRootConditionalWeakTableElementEdges[0]),
-                &pContext->rgGCBulkRootConditionalWeakTableElementEdges[0]);
-
-            pContext->iCurBulkRootConditionalWeakTableElementEdge++;
-            pContext->ClearRootConditionalWeakTableElementEdges();
-        }
-    }
-    else
-    {
-        _ASSERTE(pContext->cGcBulkRootEdges < _countof(pContext->rgGcBulkRootEdges));
-        EventStructGCBulkRootEdgeValue* pBulkRootEdgeValue = &pContext->rgGcBulkRootEdges[pContext->cGcBulkRootEdges];
-        pBulkRootEdgeValue->RootedNodeAddress = pRootedNode;
-        pBulkRootEdgeValue->GCRootKind = nRootKind;
-        pBulkRootEdgeValue->GCRootFlag = rootFlags;
-        pBulkRootEdgeValue->GCRootID = pvRootID;
-        pContext->cGcBulkRootEdges++;
-
-        // If root edge buffer is now full, empty it into ETW
-        if (pContext->cGcBulkRootEdges == _countof(pContext->rgGcBulkRootEdges))
-        {
-            FireEtwGCBulkRootEdge(
-                pContext->iCurBulkRootEdge,
-                pContext->cGcBulkRootEdges,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGcBulkRootEdges[0]),
-                &pContext->rgGcBulkRootEdges[0]);
-
-            pContext->iCurBulkRootEdge++;
-            pContext->ClearRootEdges();
-        }
-    }
-}
-
-
-//---------------------------------------------------------------------------------------
-//
-// Called during a heap walk for each object reference encountered.  Batches up the
-// corresponding node, edges, and type data for the ETW events.
-//
-// Arguments:
-//      * profilerWalkHeapContext - The shared profapi/etw context built up during the heap walk.
-//      * pObjReferenceSource - Object doing the pointing
-//      * typeID - Type of pObjReferenceSource
-//      * fDependentHandle - nonzero iff this is for a dependent handle
-//      * cRefs - Count of objects being pointed to
-//      * rgObjReferenceTargets - Array of objects being pointed to
-//
-
-// static
-void ETW::GCLog::ObjectReference(
-    ProfilerWalkHeapContext* profilerWalkHeapContext,
-    Object* pObjReferenceSource,
-    ULONGLONG typeID,
-    ULONGLONG cRefs,
-    Object** rgObjReferenceTargets)
-{
-    CONTRACTL
-    {
-        NOTHROW;
-        GC_NOTRIGGER;
-        MODE_ANY;
-
-        // LogTypeAndParametersIfNecessary can take a lock
-        CAN_TAKE_LOCK;
-    }
-    CONTRACTL_END;
-
-    EtwGcHeapDumpContext* pContext =
-        EtwGcHeapDumpContext::GetOrCreateInGCContext(&profilerWalkHeapContext->pvEtwContext);
-    if (pContext == NULL)
-        return;
-
-    //---------------------------------------------------------------------------------------
-    //    GCBulkNode events
-    //---------------------------------------------------------------------------------------
-
-    // Add Node (pObjReferenceSource) to buffer
-    _ASSERTE(pContext->cGcBulkNodeValues < _countof(pContext->rgGcBulkNodeValues));
-    EventStructGCBulkNodeValue* pBulkNodeValue = &pContext->rgGcBulkNodeValues[pContext->cGcBulkNodeValues];
-    pBulkNodeValue->Address = pObjReferenceSource;
-    pBulkNodeValue->Size = pObjReferenceSource->GetSize();
-    pBulkNodeValue->TypeID = typeID;
-    pBulkNodeValue->EdgeCount = cRefs;
-    pContext->cGcBulkNodeValues++;
-
-    // If Node buffer is now full, empty it into ETW
-    if (pContext->cGcBulkNodeValues == _countof(pContext->rgGcBulkNodeValues))
-    {
-        FireEtwGCBulkNode(
-            pContext->iCurBulkNodeEvent,
-            pContext->cGcBulkNodeValues,
-            GetClrInstanceId(),
-            sizeof(pContext->rgGcBulkNodeValues[0]),
-            &pContext->rgGcBulkNodeValues[0]);
-
-        pContext->iCurBulkNodeEvent++;
-        pContext->ClearNodes();
-    }
-
-    //---------------------------------------------------------------------------------------
-    //    BulkType events
-    //---------------------------------------------------------------------------------------
-
-    // We send type information as necessary--only for nodes, and only for nodes that we
-    // haven't already sent type info for
-    if (typeID != 0)
-    {
-        ETW::TypeSystemLog::LogTypeAndParametersIfNecessary(
-            &pContext->bulkTypeEventLogger,     // Batch up this type with others to minimize events
-            typeID,
-
-            // During heap walk, GC holds the lock for us, so we can directly enter the
-            // hash to see if the type has already been logged
-            ETW::TypeSystemLog::kTypeLogBehaviorAssumeLockAndLogIfFirstTime
-        );
-    }
-
-    //---------------------------------------------------------------------------------------
-    //    GCBulkEdge events
-    //---------------------------------------------------------------------------------------
-
-    // Add Edges (rgObjReferenceTargets) to buffer. Buffer could fill up before all edges
-    // are added (it could even fill up multiple times during this one call if there are
-    // a lot of edges), so empty Edge buffer into ETW as we go along, as many times as we
-    // need.
-
-    for (ULONGLONG i = 0; i < cRefs; i++)
-    {
-        _ASSERTE(pContext->cGcBulkEdgeValues < _countof(pContext->rgGcBulkEdgeValues));
-        EventStructGCBulkEdgeValue* pBulkEdgeValue = &pContext->rgGcBulkEdgeValues[pContext->cGcBulkEdgeValues];
-        pBulkEdgeValue->Value = rgObjReferenceTargets[i];
-        // FUTURE: ReferencingFieldID
-        pBulkEdgeValue->ReferencingFieldID = 0;
-        pContext->cGcBulkEdgeValues++;
-
-        // If Edge buffer is now full, empty it into ETW
-        if (pContext->cGcBulkEdgeValues == _countof(pContext->rgGcBulkEdgeValues))
-        {
-            FireEtwGCBulkEdge(
-                pContext->iCurBulkEdgeEvent,
-                pContext->cGcBulkEdgeValues,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGcBulkEdgeValues[0]),
-                &pContext->rgGcBulkEdgeValues[0]);
-
-            pContext->iCurBulkEdgeEvent++;
-            pContext->ClearEdges();
-        }
-    }
-}
-
-//---------------------------------------------------------------------------------------
-//
-// Called by GC at end of heap dump to give us a convenient time to flush any remaining
-// buffers of data to ETW
-//
-// Arguments:
-//      profilerWalkHeapContext - Context containing data we've batched up
-//
-
-// static
-void ETW::GCLog::EndHeapDump(ProfilerWalkHeapContext* profilerWalkHeapContext)
-{
-    LIMITED_METHOD_CONTRACT;
-
-    // If context isn't already set up for us, then we haven't been collecting any data
-    // for ETW events.
-    EtwGcHeapDumpContext* pContext = (EtwGcHeapDumpContext*)profilerWalkHeapContext->pvEtwContext;
-    if (pContext == NULL)
-        return;
-
-    // If the GC events are enabled, flush any remaining root, node, and / or edge data
-    if (RUNTIME_PROVIDER_CATEGORY_ENABLED(
-        TRACE_LEVEL_INFORMATION,
-        CLR_GCHEAPDUMP_KEYWORD))
-    {
-        if (pContext->cGcBulkRootEdges > 0)
-        {
-            FireEtwGCBulkRootEdge(
-                pContext->iCurBulkRootEdge,
-                pContext->cGcBulkRootEdges,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGcBulkRootEdges[0]),
-                &pContext->rgGcBulkRootEdges[0]);
-        }
-
-        if (pContext->cGCBulkRootConditionalWeakTableElementEdges > 0)
-        {
-            FireEtwGCBulkRootConditionalWeakTableElementEdge(
-                pContext->iCurBulkRootConditionalWeakTableElementEdge,
-                pContext->cGCBulkRootConditionalWeakTableElementEdges,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGCBulkRootConditionalWeakTableElementEdges[0]),
-                &pContext->rgGCBulkRootConditionalWeakTableElementEdges[0]);
-        }
-
-        if (pContext->cGcBulkNodeValues > 0)
-        {
-            FireEtwGCBulkNode(
-                pContext->iCurBulkNodeEvent,
-                pContext->cGcBulkNodeValues,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGcBulkNodeValues[0]),
-                &pContext->rgGcBulkNodeValues[0]);
-        }
-
-        if (pContext->cGcBulkEdgeValues > 0)
-        {
-            FireEtwGCBulkEdge(
-                pContext->iCurBulkEdgeEvent,
-                pContext->cGcBulkEdgeValues,
-                GetClrInstanceId(),
-                sizeof(pContext->rgGcBulkEdgeValues[0]),
-                &pContext->rgGcBulkEdgeValues[0]);
-        }
-    }
-
-    // Ditto for type events
-    if (RUNTIME_PROVIDER_CATEGORY_ENABLED(
-        TRACE_LEVEL_INFORMATION,
-        CLR_TYPE_KEYWORD))
-    {
-        pContext->bulkTypeEventLogger.FireBulkTypeEvent();
-        pContext->bulkTypeEventLogger.Cleanup();
-    }
-
-    // Delete any GC state built up in the context
-    profilerWalkHeapContext->pvEtwContext = NULL;
-    delete pContext;
-}
-
-#endif // FEATURE_ETW
